Renamed contracts and added use case diagrams
