+++ packages/contract/artifacts/contracts/templates/MerkleAirdropBase.sol/MerkleAirdropBase.json
@@ -1,7 +1,7 @@
 {
   "_format": "hh-sol-artifact-1",
-  "contractName": "BaseTemplate",
-  "sourceName": "contracts/templates/BaseTemplate.sol",
+  "contractName": "MerkleAirdropBase",
+  "sourceName": "contracts/templates/MerkleAirdropBase.sol",
   "abi": [
     {
       "inputs": [
@@ -191,8 +191,8 @@
       "type": "function"
     }
   ],
-  "bytecode": "0x60c060405234801561001057600080fd5b506040516101e83803806101e883398101604081905261002f91610062565b6001600160a01b0391821660a05216608052610095565b80516001600160a01b038116811461005d57600080fd5b919050565b6000806040838503121561007557600080fd5b61007e83610046565b915061008c60208401610046565b90509250929050565b60805160a0516101306100b8600039600060d80152600060b201526101306000f3fe6080604052348015600f57600080fd5b506004361060465760003560e01c80632eb4a7ab14604b5780638da5cb5b146066578063ae2e933b1460ae578063c45a01551460d4575b600080fd5b605360015481565b6040519081526020015b60405180910390f35b600054608a90610100900473ffffffffffffffffffffffffffffffffffffffff1681565b60405173ffffffffffffffffffffffffffffffffffffffff9091168152602001605d565b608a7f000000000000000000000000000000000000000000000000000000000000000081565b608a7f00000000000000000000000000000000000000000000000000000000000000008156fea2646970667358221220f378040bf99629f431782b40df9f9da3705c39a104cfdda41eb934f3bb3ae70a64736f6c63430008120033",
-  "deployedBytecode": "0x6080604052348015600f57600080fd5b506004361060465760003560e01c80632eb4a7ab14604b5780638da5cb5b146066578063ae2e933b1460ae578063c45a01551460d4575b600080fd5b605360015481565b6040519081526020015b60405180910390f35b600054608a90610100900473ffffffffffffffffffffffffffffffffffffffff1681565b60405173ffffffffffffffffffffffffffffffffffffffff9091168152602001605d565b608a7f000000000000000000000000000000000000000000000000000000000000000081565b608a7f00000000000000000000000000000000000000000000000000000000000000008156fea2646970667358221220f378040bf99629f431782b40df9f9da3705c39a104cfdda41eb934f3bb3ae70a64736f6c63430008120033",
+  "bytecode": "0x60c060405234801561001057600080fd5b506040516101e83803806101e883398101604081905261002f91610062565b6001600160a01b0391821660a05216608052610095565b80516001600160a01b038116811461005d57600080fd5b919050565b6000806040838503121561007557600080fd5b61007e83610046565b915061008c60208401610046565b90509250929050565b60805160a0516101306100b8600039600060d80152600060b201526101306000f3fe6080604052348015600f57600080fd5b506004361060465760003560e01c80632eb4a7ab14604b5780638da5cb5b146066578063ae2e933b1460ae578063c45a01551460d4575b600080fd5b605360015481565b6040519081526020015b60405180910390f35b600054608a90610100900473ffffffffffffffffffffffffffffffffffffffff1681565b60405173ffffffffffffffffffffffffffffffffffffffff9091168152602001605d565b608a7f000000000000000000000000000000000000000000000000000000000000000081565b608a7f00000000000000000000000000000000000000000000000000000000000000008156fea2646970667358221220cfaa5d67947c224e1e332f6cd2ffb6cab0f948eece8c81317cea3ace770b560264736f6c63430008120033",
+  "deployedBytecode": "0x6080604052348015600f57600080fd5b506004361060465760003560e01c80632eb4a7ab14604b5780638da5cb5b146066578063ae2e933b1460ae578063c45a01551460d4575b600080fd5b605360015481565b6040519081526020015b60405180910390f35b600054608a90610100900473ffffffffffffffffffffffffffffffffffffffff1681565b60405173ffffffffffffffffffffffffffffffffffffffff9091168152602001605d565b608a7f000000000000000000000000000000000000000000000000000000000000000081565b608a7f00000000000000000000000000000000000000000000000000000000000000008156fea2646970667358221220cfaa5d67947c224e1e332f6cd2ffb6cab0f948eece8c81317cea3ace770b560264736f6c63430008120033",
   "linkReferences": {},
   "deployedLinkReferences": {}
 }
