+++ packages/contract/contracts/templates/LinearVesting.sol
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 pragma solidity ^0.8.18;
 
-import "./BaseTemplate.sol";
+import "./MerkleAirdropBase.sol";
 
-contract MerkleAirdropLinearVesting is BaseTemplate {
+contract LinearVesting is MerkleAirdropBase {
     using SafeERC20 for IERC20;
 
     error NothingToClaim();
@@ -29,7 +29,7 @@ contract MerkleAirdropLinearVesting is BaseTemplate {
     constructor(
         address factory_,
         address feePool_
-    ) BaseTemplate(factory_, feePool_) {}
+    ) MerkleAirdropBase(factory_, feePool_) {}
 
     function initialize(
         address owner_,
