+++ packages/contract/docs/templates/LinearVesting/index.md
@@ -46,7 +46,7 @@
 constructor(
     address factory_,
     address feePool_
-) BaseTemplate(factory_, feePool_)
+) MerkleAirdropBase(factory_, feePool_)
 ```
 
 | 引数名    | 型      | 概要                         | 制約 |
